Refactor some useful common methods
