+++ IS0502Test.py
@@ -16,8 +16,6 @@ import json
 import time
 import uuid
 
-from urllib.parse import urlparse
-
 from TestResult import Test
 from GenericTest import GenericTest
 from IS05Utils import IS05Utils
@@ -92,25 +90,14 @@ class IS0502Test(GenericTest):
 
         return True, ""
 
-    def get_valid_transports(self):
-        """Identify the valid transport types for a given version of IS-05"""
-        valid_transports = ["urn:x-nmos:transport:rtp",
-                            "urn:x-nmos:transport:rtp.mcast",
-                            "urn:x-nmos:transport:rtp.ucast",
-                            "urn:x-nmos:transport:dash"]
-        api = self.apis[CONN_API_KEY]
-        if self.is05_utils.compare_api_version(api["version"], "v1.1") >= 0:
-            valid_transports.append("urn:x-nmos:transport:websocket")
-            valid_transports.append("urn:x-nmos:transport:mqtt")
-        return valid_transports
-
     def check_is04_in_is05(self, resource_type):
         """Check that each Sender or Receiver found via IS-04 has a matching entry in IS-05"""
         assert(resource_type in ["senders", "receivers"])
 
         result = True
         for is04_resource in self.is04_resources[resource_type]:
-            if is04_resource["transport"] in self.get_valid_transports():
+            valid_transports = self.is05_utils.get_valid_transports(self.apis[CONN_API_KEY]["version"])
+            if is04_resource["transport"] in valid_transports:
                 if is04_resource["id"] not in self.is05_resources[resource_type]:
                     result = False
 
@@ -132,31 +119,6 @@ class IS0502Test(GenericTest):
                 break
 
         return result
-
-    def compare_urls(self, url1, url2):
-        """Check that two URLs to a given API are sufficiently similar"""
-
-        url1_parsed = urlparse(url1.rstrip("/"))
-        url2_parsed = urlparse(url2.rstrip("/"))
-
-        comparisons = ["scheme", "hostname", "path"]
-        for attr in comparisons:
-            if getattr(url1_parsed, attr) != getattr(url2_parsed, attr):
-                return False
-
-        # Ports can be None if they are the default for the scheme
-        ports = [url1_parsed.port, url2_parsed.port]
-        comparisons = [url1_parsed, url2_parsed]
-        for index, url in enumerate(comparisons):
-            if url.port is None and url.scheme == "http":
-                ports[index] = 80
-            elif url.port is None and url.scheme == "https":
-                ports[index] = 443
-
-        if ports[0] != ports[1]:
-            return False
-
-        return True
 
     def activate_check_version(self, resource_type):
         try:
@@ -319,7 +281,7 @@ class IS0502Test(GenericTest):
                 for control in controls:
                     if control["type"] == device_type:
                         is05_devices.append(control["href"])
-                        if self.compare_urls(self.connection_url, control["href"]):
+                        if self.is05_utils.compare_urls(self.connection_url, control["href"]):
                             found_api_match = True
         except json.decoder.JSONDecodeError:
             return test.FAIL("Non-JSON response returned from Node API")
@@ -579,7 +541,8 @@ class IS0502Test(GenericTest):
         try:
             for resource_type in ["senders", "receivers"]:
                 for resource in self.is04_resources[resource_type]:
-                    if resource["transport"] not in self.get_valid_transports():
+                    valid_transports = self.is05_utils.get_valid_transports(self.apis[CONN_API_KEY]["version"])
+                    if resource["transport"] not in valid_transports:
                         continue
 
                     bindings_length = len(resource["interface_bindings"])
@@ -612,7 +575,8 @@ class IS0502Test(GenericTest):
 
         try:
             for resource in self.is04_resources["senders"]:
-                if resource["transport"] not in self.get_valid_transports():
+                valid_transports = self.is05_utils.get_valid_transports(self.apis[CONN_API_KEY]["version"])
+                if resource["transport"] not in valid_transports:
                     continue
 
                 is04_transport_file = None
